Change more integer variables from 16 to 32 bits

+++ src/parascopy/depth.py
@@ -407,8 +407,8 @@ def _predict_variance(depth_values, curr_keep_window, gc_windows, min_windows=10
 def _summarize_sample(sample, sample_window_counts, params, windows, out, res):
     n_windows = len(sample_window_counts)
     keep_window = np.ones(n_windows, dtype=np.bool_)
-    depth1 = np.zeros(n_windows, dtype=np.uint16)
-    depth2 = np.zeros(n_windows, dtype=np.uint16)
+    depth1 = np.zeros(n_windows, dtype=np.uint32)
+    depth2 = np.zeros(n_windows, dtype=np.uint32)
 
     for window_ix, counts in enumerate(sample_window_counts):
         keep_window[window_ix] = counts.passes(params)
